ml/uc_02_invoice_match/training/requirements.txt
pandas>=2.0
scikit-learn>=1.3
xgboost>=2.0
lightgbm>=4.0
mlflow>=2.8
optuna>=3.4
shap>=0.43
seaborn>=0.13
plotly>=5.18
matplotlib>=3.7
pyyaml>=6.0
joblib>=1.3
